Change type to integer

+++ models/channel.py
@@ -2,6 +2,7 @@ from elasticsearch_dsl import Boolean
 from elasticsearch_dsl import Date
 from elasticsearch_dsl import Double
 from elasticsearch_dsl import Float
+from elasticsearch_dsl import Integer
 from elasticsearch_dsl import Keyword
 from elasticsearch_dsl import Long
 from elasticsearch_dsl import Object
@@ -174,7 +175,7 @@ class ChannelSectionBrandSafety(BaseInnerDoc):
     categories = Object()
     rescore = Boolean()  # Flag used if should be rescored by brand safety script
     limbo_status = Boolean() # Flag used if vetting should be reviewed
-    pre_limbo_score = Long()  # Brand safety script score
+    pre_limbo_score = Integer()  # Brand safety script score
 
 
 class ChannelSectionAuth(BaseInnerDoc):
