+++ models/video.py
@@ -3,6 +3,7 @@ from elasticsearch_dsl import Date
 from elasticsearch_dsl import Double
 from elasticsearch_dsl import Float
 from elasticsearch_dsl import InnerDoc
+from elasticsearch_dsl import Integer
 from elasticsearch_dsl import Keyword
 from elasticsearch_dsl import Long
 from elasticsearch_dsl import Object
@@ -183,7 +184,7 @@ class VideoSectionBrandSafety(BaseInnerDoc):
     categories = Object()
     rescore = Boolean()  # Flag used if should be rescored by brand safety script
     limbo_status = Boolean()  # Flag used if vetting should be reviewed
-    pre_limbo_score = Long()  # Brand safety script score
+    pre_limbo_score = Integer()  # Brand safety script score
 
 
 class VideoSectionTaskUsData(BaseInnerDoc):
